finished main unit testing -> test_main.py with test functions and test_main_utils with test assets.
