+++ backend/requirements.txt
@@ -1,20 +1,30 @@
 aiomysql==0.2.0
+aiosqlite==0.21.0
 annotated-types==0.7.0
 anyio==4.8.0
 bcrypt==4.3.0
+certifi==2025.1.31
 cffi==1.17.1
 click==8.1.8
 cryptography==44.0.2
 fastapi==0.115.11
 greenlet==3.1.1
 h11==0.14.0
+httpcore==1.0.7
+httpx==0.28.1
 idna==3.10
+iniconfig==2.1.0
+jwt==1.3.1
 mysql-connector-python==9.2.0
+packaging==24.2
+pluggy==1.5.0
 pycparser==2.22
 pydantic==2.10.6
 pydantic_core==2.27.2
 PyJWT==2.10.1
 PyMySQL==1.1.1
+pytest==8.3.5
+pytest-asyncio==0.26.0
 python-multipart==0.0.20
 sniffio==1.3.1
 SQLAlchemy==2.0.40
