+++ requirements.txt
@@ -20,3 +20,5 @@ typing_extensions==4.12.2
 uvicorn==0.34.0
 cryptography~=44.0.2
 jwt~=1.3.1
+PyJWT~=2.10.1
+pytest~=8.3.5
